# backend/calyb-connector.openapi.yaml
openapi: 3.1.0
info:
  title: Support Ticket Dashboard — Bulk API
  description: >
    Bulk data endpoint for the Calyb connector and data view agents.
    This API is NOT used by the frontend dashboard — the frontend uses the
    separate paginated /tickets/page endpoint instead.

    Authentication: Bearer token. Use key `calyb-key-002` for the Calyb
    connector or `agent-key-003` for data view agents.
  version: 2.0.0

servers:
  - url: http://localhost:8000
    description: Local development

security:
  - BearerAuth: []

paths:
  /tickets:
    get:
      summary: Bulk ticket fetch
      description: >
        Returns all tickets matching the supplied filters in a single response.
        All filter parameters are optional and combined with AND logic.
        No pagination — use `created_after` / `created_before` to slice large
        date ranges if needed.
      operationId: getBulkTickets
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by ticket status.
          schema:
            type: string
            enum: [open, acknowledged, in_progress, waiting_on_info, resolved, closed]
          example: open

        - name: priority
          in: query
          required: false
          description: Filter by priority level.
          schema:
            type: string
            enum: [low, medium, high, critical]
          example: critical

        - name: category
          in: query
          required: false
          description: Filter by ticket category.
          schema:
            type: string
            enum: [bug, feature, billing, access, performance]
          example: billing

        - name: assigned_to
          in: query
          required: false
          description: Exact match on developer name.
          schema:
            type: string
          example: "Alice Chen"

        - name: unassigned
          in: query
          required: false
          description: >
            When true, return only tickets where `assigned_to` is null.
            Mutually exclusive with `assigned_to`.
          schema:
            type: boolean
          example: true

        - name: created_after
          in: query
          required: false
          description: >
            Return tickets created at or after this datetime.
            ISO 8601 format, e.g. `2025-03-01T00:00:00+00:00`.
          schema:
            type: string
            format: date-time
          example: "2025-03-01T00:00:00+00:00"

        - name: created_before
          in: query
          required: false
          description: >
            Return tickets created at or before this datetime.
            ISO 8601 format, e.g. `2025-03-31T23:59:59+00:00`.
          schema:
            type: string
            format: date-time
          example: "2025-03-31T23:59:59+00:00"

        - name: tag
          in: query
          required: false
          description: Return only tickets whose `tags` list contains this value.
          schema:
            type: string
          example: urgent

        - name: age_gt
          in: query
          required: false
          description: Return only tickets older than N hours (based on `created_at`).
          schema:
            type: integer
          example: 48

      responses:
        "200":
          description: Matching tickets returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkResponse"
        "401":
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
              example:
                detail: "Invalid API key"
        "422":
          description: Invalid query parameter value.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"

  /tickets/{ticket_id}:
    get:
      summary: Get single ticket
      operationId: getTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "200":
          description: Ticket found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Ticket not found.

  /tickets/{ticket_id}/assign:
    post:
      summary: Assign a ticket to a developer
      operationId: assignTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assigned_to]
              properties:
                assigned_to:
                  type: string
                  example: "Alice Chen"
      responses:
        "200":
          description: Ticket assigned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Ticket not found.

  /tickets/{ticket_id}/transfer:
    post:
      summary: Transfer a ticket to a different developer
      operationId: transferTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [developer, reason]
              properties:
                developer:
                  type: string
                  example: "Bob Patel"
                reason:
                  type: string
                  example: "Original assignee on leave"
      responses:
        "200":
          description: Ticket transferred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          description: Precondition failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreconditionError"
        "422":
          description: Missing required reason field.

  /tickets/{ticket_id}/escalate:
    post:
      summary: Escalate ticket priority by one level
      operationId: escalateTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "200":
          description: Ticket escalated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          description: Precondition failed (invalid status or already critical).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreconditionError"

  /tickets/{ticket_id}/acknowledge:
    post:
      summary: Developer acknowledges a ticket
      operationId: acknowledgeTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [developer]
              properties:
                developer:
                  type: string
                  example: "Alice Chen"
      responses:
        "200":
          description: Ticket acknowledged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          description: Precondition failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreconditionError"

  /tickets/{ticket_id}/comment:
    post:
      summary: Add a comment to a ticket
      operationId: addComment
      parameters:
        - $ref: "#/components/parameters/TicketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [author, text]
              properties:
                author:
                  type: string
                  example: "Alice Chen"
                text:
                  type: string
                  example: "I can reproduce this on staging."
      responses:
        "200":
          description: Comment added.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"

  /tickets/{ticket_id}/reopen:
    post:
      summary: Reopen a resolved, closed, or waiting_on_info ticket
      operationId: reopenTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "200":
          description: Ticket reopened. Assignee cleared, status set to open.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          description: Precondition failed (invalid source status).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreconditionError"

  /tickets/{ticket_id}/request-info:
    post:
      summary: Set ticket to waiting_on_info state
      operationId: requestInfo
      parameters:
        - $ref: "#/components/parameters/TicketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [note]
              properties:
                note:
                  type: string
                  example: "Need customer to confirm their browser version."
      responses:
        "200":
          description: Ticket set to waiting_on_info.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          description: Precondition failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreconditionError"

  /tickets/bulk-assign:
    post:
      summary: Assign multiple tickets to a single developer
      operationId: bulkAssign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ticket_ids, developer]
              properties:
                ticket_ids:
                  type: array
                  items:
                    type: string
                  example: ["TKT-0001", "TKT-0002"]
                developer:
                  type: string
                  example: "Alice Chen"
      responses:
        "200":
          description: Bulk assignment result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  assigned:
                    type: array
                    items:
                      type: string
                  skipped:
                    type: array
                    items:
                      type: string
                  developer:
                    type: string

  /tickets/{ticket_id}/history:
    get:
      summary: Get full audit log for a ticket
      operationId: getTicketHistory
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "200":
          description: History entries returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket_id:
                    type: string
                  history:
                    type: array
                    items:
                      $ref: "#/components/schemas/HistoryEntry"

  /developers:
    get:
      summary: List all developers with open ticket count and specialisations
      operationId: getDevelopers
      responses:
        "200":
          description: Developer list returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  developers:
                    type: array
                    items:
                      $ref: "#/components/schemas/Developer"

  /tickets/{ticket_id}/status:
    post:
      summary: Update ticket status (state-machine enforced)
      operationId: updateTicketStatus
      parameters:
        - $ref: "#/components/parameters/TicketId"
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum: [open, acknowledged, in_progress, waiting_on_info, resolved, closed]
      responses:
        "200":
          description: Status updated.
        "400":
          description: Invalid transition.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreconditionError"

  /calyb/planner-test:
    post:
      summary: Planner test harness endpoint
      description: >
        Test-only endpoint for validating planner accuracy. Accepts a plan
        and returns evaluation results.
      operationId: plannerTest
      x-test-only: true
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Test results returned.

  /health:
    get:
      summary: Health check
      description: Returns server status and number of loaded tickets. No auth required.
      operationId: getHealth
      security: []
      responses:
        "200":
          description: Server is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  tickets_loaded:
                    type: integer
                    example: 200
                required: [status, tickets_loaded]

  /summary:
    get:
      summary: Dashboard summary statistics
      operationId: getSummary
      responses:
        "200":
          description: Summary statistics returned.

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Dummy API keys for testing:
        - `calyb-key-002` — Calyb Connector (reader)
        - `agent-key-003` — Data View Agent (agent — read + write + planner-test)
        - `dashboard-key-001` — Frontend Dashboard (admin, not for connector use)

  parameters:
    TicketId:
      name: ticket_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^TKT-\\d{4}$"
      example: "TKT-0042"

  schemas:
    Ticket:
      type: object
      required:
        - id
        - title
        - description
        - status
        - priority
        - category
        - assigned_to
        - created_at
        - updated_at
        - created_by
        - tags
      properties:
        id:
          type: string
          pattern: "^TKT-\\d{4}$"
          example: "TKT-0042"
        title:
          type: string
          example: "HTTP 500 on checkout when cart has 50+ items"
        description:
          type: string
          example: "When a user adds more than 50 items to their cart..."
        status:
          type: string
          enum: [open, acknowledged, in_progress, waiting_on_info, resolved, closed]
          example: open
        priority:
          type: string
          enum: [low, medium, high, critical]
          example: critical
        category:
          type: string
          enum: [bug, feature, billing, access, performance]
          example: bug
        assigned_to:
          type: ["string", "null"]
          example: "Alice Chen"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          example: "Bob Patel"
        tags:
          type: array
          items:
            type: string
          example: ["urgent", "auth"]
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        history:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"

    Comment:
      type: object
      required: [author, text, timestamp]
      properties:
        author:
          type: string
          example: "Alice Chen"
        text:
          type: string
          example: "I can reproduce this on staging."
        timestamp:
          type: string
          format: date-time

    HistoryEntry:
      type: object
      required: [action, actor, timestamp]
      properties:
        action:
          type: string
          example: "status_change"
        actor:
          type: string
          example: "Alice Chen"
        from_status:
          type: ["string", "null"]
        to_status:
          type: ["string", "null"]
        timestamp:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]

    Developer:
      type: object
      required: [name, open_ticket_count, specialisations]
      properties:
        name:
          type: string
          example: "Alice Chen"
        open_ticket_count:
          type: integer
          example: 12
        specialisations:
          type: array
          items:
            type: string
          example: ["backend", "auth"]

    BulkResponse:
      type: object
      required: [total, tickets]
      properties:
        total:
          type: integer
          example: 11
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"

    PreconditionError:
      type: object
      required: [error, reason, detail]
      properties:
        error:
          type: string
          example: "precondition_failed"
        reason:
          type: string
          example: "invalid_transition"
        detail:
          type: string
          example: "Cannot transition from 'closed' to 'in_progress'."

    ErrorDetail:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "Invalid API key"

    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"

    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string